requirements.txt
torch>=1.11
torchvision
timm
open_clip_torch
opencv-python
matplotlib
numpy
requests
einops
